requirements.txt
mcp[cli]>=1.26.0
httpx>=0.27
python-dotenv>=1.0
